final version. correct

+++ CountInversions/TestCase.cs
@@ -10,6 +10,6 @@ namespace CountInversions
     {
         public string FilePath { get; set; }
         public int[] Input { get; set; }
-        public int InversionsCount { get; set; }
+        public long InversionsCount { get; set; }
     }
 }
